Fix bugs with new custom view.

+++ app/src/main/java/com/intelliviz/quakereport/Earthquake.kt
@@ -1,5 +1,5 @@
 package com.intelliviz.quakereport
 
-data class Earthquake(val magnitude: Double,
+data class Earthquake(val magnitude: Float,
                       val distance: String, val city: String, val date: String,
                       val time: String, val url: String)
