+++ app/src/main/java/com/intelliviz/quakereport/EarthquakeInfo.kt
@@ -1,3 +1,3 @@
 package com.intelliviz.quakereport
 
-data class EarthquakeInfo(val year: Int, val magnitude: Int, val count: Int)
+data class EarthquakeInfo(val year: Int, val magnitude: Float, val count: Int)
